requirements.txt
anitopy
swibots
python-decouple